# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Tracker Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Job Applications
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List job applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update a job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete a job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Analytics
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        location:
          type: string
        notes:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        location:
          type: string
        notes:
          type: string

    # Applications
    Application:
      type: object
      required: [id, jobTitle, status, createdAt]
      properties:
        id:
          type: integer
        jobTitle:
          type: string
        status:
          type: string
          enum: [applied, screening, interview, offer, rejected, withdrawn]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        jobUrl:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        appliedAt:
          type: ["string", "null"]
        nextFollowUpAt:
          type: ["string", "null"]
        interviewStage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApplicationInput:
      type: object
      required: [jobTitle]
      properties:
        jobTitle:
          type: string
        companyId:
          type: integer
        jobUrl:
          type: string
        salary:
          type: string
        location:
          type: string
        notes:
          type: string
        appliedAt:
          type: string
        nextFollowUpAt:
          type: string
        status:
          type: string
          enum: [applied, screening, interview, offer, rejected, withdrawn]
    ApplicationUpdate:
      type: object
      properties:
        jobTitle:
          type: string
        companyId:
          type: ["integer", "null"]
        jobUrl:
          type: string
        salary:
          type: string
        location:
          type: string
        notes:
          type: string
        appliedAt:
          type: string
        nextFollowUpAt:
          type: string
        interviewStage:
          type: string
        status:
          type: string
          enum: [applied, screening, interview, offer, rejected, withdrawn]
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [applied, screening, interview, offer, rejected, withdrawn]
        interviewStage:
          type: string

    # Analytics
...
    Notification:
      type: object
      required: [id, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        applicationId:
          type: ["integer", "null"]
        createdAt:
          type: string

    # AI
...